refactor(trading): relocate donation views
