+++ bot/features/trading/views/__init__.py
@@ -6,6 +6,11 @@ from .common import (
 from .common import (
     format_coins as format_coins,
 )
+from .donate import (
+    DonateAmountModal as DonateAmountModal,
+    DonateSelect as DonateSelect,
+    DonateView as DonateView,
+)
 from .inventory import (
     InventorySelect as InventorySelect,
 )
@@ -21,6 +26,9 @@ from .shop import (
 
 __all__ = [
     "CloseButton",
+    "DonateAmountModal",
+    "DonateSelect",
+    "DonateView",
     "InventorySelect",
     "InventoryView",
     "ShopSelect",
